Rename lru_cache pkg to cache
